[기능 수정] 게시글 컴포넌트 세분화
